Fix broadcom squashfs images:
We switched over to appending the jffs2 eof mark to the squashfs images, 
but since the squashfs is not always aligned to eraseblocksize, the eof
mark landed in the wrong place. This commit adds an extra flag to the
trx utility that can append extra data to a partition with alignment.
This is used to place the jffs2 eof mark at the right offset.



git-svn-id: https://svn.openwrt.org/openwrt/trunk@7253 3c298f89-4303-0410-b956-a3cf2f4a3e73