don't specify "CC=<nothing>" on kernel build command line
If KERNEL_CC isn't set, we end up with a "CC=" on the kernel build
command-line. We don't always need CC, as the CROSS_COMPILE flag does
the job instead. In fact, specifying CC messes up the build when we're
using a biarch compiler.

This change doesn't specify CC= if the KERNEL_CC variable is empty.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>


git-svn-id: https://svn.openwrt.org/openwrt/trunk@9408 3c298f89-4303-0410-b956-a3cf2f4a3e73