Allow targets to specify extra initramfs source files
The CONFIG_INITRAMFS_SOURCE Kconfig variable can be a space-separated
list of source files (or directories). This allows a platform to
add extra components to the initramfs image, by defining the
INITRAMFS_EXTRA_FILES make var.

By default, we add a simple initramfs extra file for the generic-2.6
platform, which specifies a few device nodes.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>


git-svn-id: https://svn.openwrt.org/openwrt/trunk@9410 3c298f89-4303-0410-b956-a3cf2f4a3e73