DG834G Power LED fix
This patch enables the new "default-on" trigger for the AR7 target and uses it to ensure the "power" LED on the DG834G is initialised in the ON state.

Previously this LED would turn off when the driver loaded and stay off until turned back on from userspace. With this patch it remains on throughout the boot process, but can still be controlled from userspace if required.

Signed-off-by: Nick Forbes <nick.forbes@incepta.com>


git-svn-id: https://svn.openwrt.org/openwrt/trunk@10349 3c298f89-4303-0410-b956-a3cf2f4a3e73