bcm963xx: add cfe image tagging utility
This replaces the proprietary bcmImageBuilder program from the Broadcom source
drops. It basically adds a 256 bytes header in front of the kernel + rootfs
which contains a few text signatures, the locations of the data as well as the
checksums of the data and the tag. It also adds a 12 bytes header in front of
the LZMA kernel which contains the load address, kernel entry and the size of
the compressed LZMA data.

Signed-off-by: Axel Gembe <ago@bastart.eu.org>


git-svn-id: https://svn.openwrt.org/openwrt/trunk@11170 3c298f89-4303-0410-b956-a3cf2f4a3e73