(4/6) bcm57xx: switch-core.c/switch-robo.c check for port already registered
This patch prevents switch-robo.c from attempting robo_probe on a port
that is already registered.  robo_probe will adjust kernel reference counts
if it detects a switch on the port.  If this patch wasn't applied, the
wrt350n would hang on reboot, waiting for the network driver reference count
to reach zero indefinitely.

Signed-off-by: Ben Pfountz <netprince (at) vt (dot) edu>


git-svn-id: https://svn.openwrt.org/openwrt/trunk@11469 3c298f89-4303-0410-b956-a3cf2f4a3e73